Create basic_network_sniffer_task1.py
This Python code utilizes the Scapy library to capture and analyze network packets. It defines a function packet_callback that is called for each packet captured. Within this function, it checks if the packet has an IP layer and if so, it prints out the source and destination IP addresses. The code then specifies the network interface (interface) to sniff packets on, sets a timeout of 10 seconds, and captures packets using the sniff function, calling the packet_callback for each packet. Finally, it prints out the number of packets captured. Note that the network interface name ('Wi-Fi') might need to be replaced with the actual name of the Wi-Fi interface on the system running the code.
Certainly! In this code, the network interface variable interface is set to 'Wi-Fi'. Users would need to replace 'Wi-Fi' with the actual name of their Wi-Fi interface. This flexibility allows the code to adapt to different network setups, ensuring it works on a variety of systems. So, users can modify the interface variable according to their specific network configuration before running the code.